Rkuris/move path iter item (#668)
