+++ storage/src/lib.rs
@@ -20,7 +20,7 @@ mod trie_hash;
 
 // re-export these so callers don't need to know where they are
 pub use linear::{LinearStoreParent, ReadLinearStore, WriteLinearStore};
-pub use node::{path::NibblesIterator, path::Path, BranchNode, LeafNode, Node};
+pub use node::{path::NibblesIterator, path::Path, BranchNode, LeafNode, Node, PathIterItem};
 pub use nodestore::{LinearAddress, NodeStore, UpdateError};
 
 pub use linear::proposed::{ProposedImmutable, ProposedMutable};
